Let us move the fallback to its own file so as not to mess  up with stage 2 too much.
